+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### SliceRecon
 - Expose reconstructor to the Python bindings
 - Allow setting a custom filter using an array instead of a file
+- Add `mute` functionality to logger
 
 ## [1.1.0] - 2020-27-03